requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.2
frozendict==2.4.4
h11==0.14.0
html5lib==1.1
idna==3.8
lxml==5.3.0
multitasking==0.0.11
numpy==2.1.0
pandas==2.2.2
passlib==1.7.4
peewee==3.17.6
platformdirs==4.2.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
webencodings==0.5.1
yfinance==0.2.43